moved widgets to a widget folder
