+++ lib/widgets/portfolio_widget.dart
@@ -1,0 +1,56 @@
+import 'package:flutter/material.dart';
+
+
+class PortfolioWidget extends StatelessWidget {
+  const PortfolioWidget({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    List<String> portfolioimages = [
+      'assets/samplepictures/fix1.jpg',
+      'assets/samplepictures/fix2.jpg',
+      'assets/samplepictures/fix3.jpg',
+      'assets/samplepictures/fix1.jpg',
+      'assets/samplepictures/fix2.jpg',
+      'assets/samplepictures/fix3.jpg',
+      'assets/samplepictures/fix1.jpg',
+      'assets/samplepictures/fix2.jpg',
+      'assets/samplepictures/fix3.jpg',
+      'assets/samplepictures/fix1.jpg',
+      'assets/samplepictures/fix2.jpg',
+      'assets/samplepictures/fix3.jpg',
+    ]; 
+    return Container(
+        decoration: BoxDecoration(
+          border: Border.all(color: const Color(0xFFFF5C00)),
+          borderRadius: const BorderRadius.all(Radius.circular(10)),
+         ),
+        padding: const EdgeInsets.all(20.0),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            const Text("Portfolio\n", 
+              style: TextStyle(
+                fontWeight: FontWeight.bold,
+                fontSize: 24,
+                color: Color.fromARGB(255, 255, 18, 227),
+              )),
+            GridView.builder(
+              shrinkWrap: true,
+              gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
+                crossAxisCount: 3,
+                crossAxisSpacing: 5,
+                mainAxisSpacing: 5,
+                ),
+              itemCount: portfolioimages.length, //number of items in portfolioimages
+              itemBuilder: (context, index) {
+               return Image.asset(portfolioimages[index],fit: BoxFit.cover);
+              },
+            )
+          ]
+        )
+    );
+  }
+}
